Logic/add order details position logic

+++ Firma/Models/BusinessLogic/Calculations/Calculate.cs
@@ -34,7 +34,7 @@
             {
                 return netPrice;
             }
-            return ((netPrice + (netPrice * (vat / 100))) * quantity);
+            return ((netPrice + (netPrice * (vat / 100.00m))) * (decimal)quantity);
         }
     }
 }
